+++ test/file.spec.js
@@ -5,17 +5,19 @@ describe('Core input method: file', function() {
   it('Should evaluate all urls', async function() {
     this.timeout(0);
 
+    const qualweb = new QualWeb({ adBlock: true, stealth: true });
+
     const options = { 
-      file: './test/urls.txt',
-      maxParallelEvaluations: 10
+      file: 'urls.txt',
+      'wcag-techniques': {
+        exclude: ['QW-WCAG-T16']
+      }
     };
 
-    const qualweb = new QualWeb({ stealth: true, adBlock: true });
-
-    await qualweb.start();
+    await qualweb.start({ maxConcurrency: 5, monitor: true }, { args: ['--no-sandbox', '--ignore-certificate-errors']});
     const reports = await qualweb.evaluate(options);
     await qualweb.stop();
     
-    expect(Object.keys(reports).length).to.be.equal(10);
+    expect(Object.keys(reports).length).to.be.equal(25);
   });
 });